# Conflicts:
#	esphome/core/helpers.cpp

+++ esphome/core/helpers.cpp
@@ -400,8 +400,8 @@ std::string format_hex_pretty(const std::vector<uint16_t> &data) { return format
 std::string format_bin(const uint8_t *data, size_t length) {
   std::string result;
   result.resize(length * 8);
-  for (int byte_idx = 0; byte_idx < length; byte_idx++) {
-    for (int bit_idx = 0; bit_idx < 8; bit_idx++) {
+  for (size_t byte_idx = 0; byte_idx < length; byte_idx++) {
+    for (size_t bit_idx = 0; bit_idx < 8; bit_idx++) {
       result[byte_idx * 8 + bit_idx] = ((data[byte_idx] >> (7 - bit_idx)) & 1) + '0';
     }
   }
